MCP_Folder/gmailmcp_project1/pyproject.toml
[project]
name = "smart-gmail-mcp"
version = "0.1.0"
dependencies = [
    "fastmcp",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "pydantic",
    "python-dotenv",
    "loguru",
    "tenacity",
]
